Adding object selection

+++ src/main/java/com/Ecommerce/acme/model/Product.java
@@ -9,9 +9,6 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
-import org.hibernate.annotations.Fetch;
-import org.hibernate.annotations.FetchMode;
-
 import lombok.Data;
 
 @Entity
@@ -23,7 +20,7 @@ public class Product {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private int id_product;
 	private String name;
-	private float unit_price;
+	private int unit_price;
 	private String picture;
 	
 	@ManyToOne(fetch = FetchType.LAZY)
